+++ tests/Functional/RobotsParser/RobotsParserTest.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace Tests\Functional\RobotsParser;
+
+use Sobak\Scrawler\Block\ResultWriter\InMemoryResultWriter;
+use Sobak\Scrawler\Block\RobotsParser\DefaultRobotsParser;
+use Sobak\Scrawler\Configuration\ObjectConfiguration;
+use Sobak\Scrawler\Matcher\CssSelectorListMatcher;
+use Sobak\Scrawler\Matcher\CssSelectorTextMatcher;
+use Sobak\Scrawler\Scrawler;
+use Tests\Functional\ServerBasedTest;
+use Tests\Utils\BasicConfigurationProvider;
+use Tests\Utils\SimpleMatchEntity;
+
+class RobotsParserTest extends ServerBasedTest
+{
+    public function testRobotsTxtParser(): void
+    {
+        $config = BasicConfigurationProvider::getConfiguration()
+            ->setBaseUrl(ServerBasedTest::getHostUrl())
+            ->setRobotsParser(new DefaultRobotsParser())
+            ->addObjectDefinition('test', new CssSelectorListMatcher('body'), function (ObjectConfiguration $object) {
+                $object
+                    ->addFieldDefinition('match', new CssSelectorTextMatcher('span.match'))
+                    ->addEntityMapping(SimpleMatchEntity::class)
+                    ->addResultWriter(SimpleMatchEntity::class, new InMemoryResultWriter())
+                ;
+            })
+        ;
+
+        $scrawler = new Scrawler($config, __DIR__ . '/output');
+        $scrawler->run();
+
+        $this->assertEmpty(InMemoryResultWriter::$results);
+    }
+}